Created decorator for LTI
